second

+++ MyRandomCurisity/CNum6MoreThenOneInteger.cpp
@@ -7,7 +7,7 @@ int main() {
     int n;
     cin >> n; 
 
-    vector<int> arr(n);
+    vector<long long> arr(n);
     for (int i = 0; i < n; i++) {
         cin >> arr[i];
     }
